# perf: Cache locality optimization for airspace clipping - 57% improvement